2020.01.06: Google Stock Model

+++ model.py
@@ -79,7 +79,7 @@ class KerasPredictor:
 
         lstm.compile(optimizer='adam', loss='mean_squared_error')
         lstm.fit(self.training_input, self.training_output, epochs=iterations, batch_size=batch_size) # train each time series 1000 times
-        lstm.fit(self.validation_input, self.validation_output, epochs=iterations/10, batch_size=batch_size) 
+        lstm.fit(self.validation_input, self.validation_output, epochs=int(iterations/10), batch_size=batch_size)
 
         # save the model
         model_name = self.stock.stock_name() + "_model.h5"
